fixed bugs in render and added blending
perspective matrix calculations had to be fixed to correctly render models. other small fixes in matrix class to correct matrix multiplication. added in pose blending and smooth transitioning. additive blending to be added soon

+++ AnimationProject/Clip.cpp
@@ -91,7 +91,7 @@ void Clip::RecalculateDuration()
 	mEndTime = 0.0f;
 	bool startSet = false;
 	bool endSet = false;
-	unsigned int trackSize = mTracks.size();
+	unsigned int trackSize = (unsigned int)mTracks.size();
 	for (unsigned int i = 0; i < trackSize; ++i)
 	{
 		if (mTracks[i].IsValid())
